+++ Views/Pages/PlaylistsPage.xaml.cs
@@ -1,9 +1,0 @@
-using System.Windows.Controls;
-
-namespace SandstormModLauncher.Views.Pages
-{
-    public partial class PlaylistsPage : UserControl
-    {
-        public PlaylistsPage() { InitializeComponent(); }
-    }
-}